backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
firebase-admin==6.6.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
numpy==2.2.0
scikit-learn==1.6.0
pandas==2.2.3
python-multipart==0.0.20
